Split worktree terminal state helpers
